+++ mentions.js
@@ -1,0 +1,5 @@
+const MentionsIndex = require('./indexes/mentions')
+
+exports.init = function (sbot, config) {
+  sbot.db.registerIndex(MentionsIndex)
+}
