+++ operators/index.js
@@ -6,23 +6,47 @@ const {
   seekChannel,
   seekRoot,
   seekPrivate,
+  seekVoteLink,
 } = require('../seekers')
 const { equal } = jitdbOperators
 
 function key(value) {
-  return equal(seekKey, value, { indexType: 'key', prefix: 32 })
+  return equal(seekKey, value, {
+    prefix: 32,
+    indexType: 'key',
+  })
 }
 
 function type(value) {
-  return equal(seekType, value, { indexType: 'value_content_type' })
+  return equal(seekType, value, {
+    indexType: 'value_content_type',
+  })
 }
 
 function author(value) {
-  return equal(seekAuthor, value, { indexType: 'value_author' })
+  return equal(seekAuthor, value, {
+    indexType: 'value_author',
+  })
 }
 
 function channel(value) {
-  return equal(seekChannel, value, { indexType: 'value_content_channel' })
+  return equal(seekChannel, value, {
+    indexType: 'value_content_channel',
+  })
+}
+
+function votesFor(msgKey) {
+  return equal(seekVoteLink, msgKey, {
+    prefix: 32,
+    indexType: 'value_content_vote_link',
+  })
+}
+
+function hasRoot(msgKey) {
+  return equal(seekRoot, msgKey, {
+    prefix: 32,
+    indexType: 'value_content_root',
+  })
 }
 
 function isRoot() {
@@ -32,8 +56,7 @@ function isRoot() {
   })
 }
 
-let bTrue = Buffer.alloc(1)
-bTrue[0] = 1
+const bTrue = Buffer.alloc(1, 1)
 function isPrivate() {
   return equal(seekPrivate, bTrue, { indexType: 'meta_private' })
 }
@@ -43,6 +66,8 @@ module.exports = Object.assign({}, jitdbOperators, {
   author,
   channel,
   key,
+  votesFor,
+  hasRoot,
   isRoot,
   isPrivate,
 })
