--- social.js
@@ -1,5 +1,0 @@
-const SocialIndex = require('./indexes/social')
-
-exports.init = function (sbot, config) {
-  sbot.db.registerIndex(SocialIndex)
-}
