+++ test/operators.js
@@ -3,6 +3,7 @@ const ssbKeys = require('ssb-keys')
 const path = require('path')
 const rimraf = require('rimraf')
 const mkdirp = require('mkdirp')
+const pull = require('pull-stream')
 const DB = require('../db')
 const {
   and,
@@ -11,6 +12,10 @@ const {
   toCallback,
   author,
   isRoot,
+  votesFor,
+  hasRoot,
+  live,
+  toPullStream,
 } = require('../operators')
 
 const dir = '/tmp/ssb-db2-operators'
@@ -45,7 +50,7 @@ test('execute and(type("post"), author(me))', (t) => {
   })
 })
 
-test('execute and(type("post"), isPrivate)', (t) => {
+test('execute and(type("post"), isPrivate())', (t) => {
   let content = { type: 'post', text: 'super secret', recps: [keys.id] }
   content = ssbKeys.box(
     content,
@@ -91,5 +96,176 @@ test('execute isRoot()', (t) => {
         })
       }
     )
+  })
+})
+
+test('execute hasRoot(msgkey)', (t) => {
+  const post = { type: 'post', text: 'Testing!' }
+  const post2 = { type: 'post', text: 'Testing 2!' }
+
+  db.publish(post, (err, postMsg) => {
+    t.error(err, 'no err')
+
+    db.publish(post2, (err) => {
+      t.error(err, 'no err')
+
+      const threadMsg1 = { type: 'post', text: 'reply', root: postMsg.key }
+
+      db.publish(threadMsg1, (err) => {
+        t.error(err, 'no err')
+
+        db.onDrain(() => {
+          db.query(
+            and(hasRoot(postMsg.key)),
+            toCallback((err, results) => {
+              t.error(err, 'no err')
+              t.equal(results.length, 1)
+              t.equal(results[0].value.content.text, threadMsg1.text)
+              t.end()
+            })
+          )
+        })
+      })
+    })
+  })
+})
+
+test('hasRoot() outputs encrypted replies too', (t) => {
+  const post = { type: 'post', text: 'Testing!' }
+  const post2 = { type: 'post', text: 'Testing 2!' }
+
+  db.publish(post, (err, postMsg) => {
+    t.error(err, 'no err')
+
+    db.publish(post2, (err) => {
+      t.error(err, 'no err')
+
+      let threadMsg1 = {
+        type: 'post',
+        text: 'encrypted reply',
+        root: postMsg.key,
+        recps: [keys.id],
+      }
+      threadMsg1 = ssbKeys.box(
+        threadMsg1,
+        threadMsg1.recps.map((x) => x.substr(1))
+      )
+
+      db.publish(threadMsg1, (err, privMsg) => {
+        t.error(err, 'no err')
+
+        db.onDrain(() => {
+          db.query(
+            and(hasRoot(postMsg.key)),
+            toCallback((err, results) => {
+              t.error(err, 'no err')
+              t.equal(results.length, 1)
+              t.equal(results[0].value.content.text, 'encrypted reply')
+              t.end()
+            })
+          )
+        })
+      })
+    })
+  })
+})
+
+test('execute votesFor(msgkey)', (t) => {
+  const post = { type: 'post', text: 'Testing!' }
+
+  db.publish(post, (err, postMsg) => {
+    t.error(err, 'no err')
+
+    const voteMsg1 = {
+      type: 'vote',
+      vote: {
+        link: postMsg.key,
+        value: 1,
+        expression: '❤',
+      },
+    }
+
+    const voteMsg2 = {
+      type: 'vote',
+      vote: {
+        link: postMsg.key,
+        value: -1,
+        expression: '❤',
+      },
+    }
+
+    db.publish(voteMsg1, (err, v1) => {
+      t.error(err, 'no err')
+
+      db.publish(voteMsg2, (err, v2) => {
+        t.error(err, 'no err')
+
+        db.onDrain(() => {
+          db.query(
+            and(votesFor(postMsg.key)),
+            toCallback((err, results) => {
+              t.error(err, 'no err')
+              t.equal(results.length, 2)
+              t.equal(results[0].key, v1.key)
+              t.equal(results[1].key, v2.key)
+              t.end()
+            })
+          )
+        })
+      })
+    })
+  })
+})
+
+test('live votesFor', (t) => {
+  const post = { type: 'post', text: 'Testing!' }
+
+  db.publish(post, (err, postMsg) => {
+    t.error(err, 'no err')
+
+    const voteMsg1 = {
+      type: 'vote',
+      vote: {
+        link: postMsg.key,
+        value: 1,
+        expression: '❤',
+      },
+    }
+
+    const voteMsg2 = {
+      type: 'vote',
+      vote: {
+        link: postMsg.key,
+        value: -1,
+        expression: '❤',
+      },
+    }
+
+    db.publish(voteMsg1, (err, v1) => {
+      t.error(err, 'no err')
+
+      db.onDrain(() => {
+        let i = 0
+        pull(
+          db.query(
+            and(votesFor(postMsg.key)),
+            live(),
+            toPullStream(),
+            pull.drain((result) => {
+              if (i++ == 0) {
+                t.equal(result.key, v1.key)
+
+                db.publish(voteMsg2, (err, v2) => {
+                  t.error(err, 'no err')
+                })
+              } else {
+                t.equal(result.value.content.vote.value, -1)
+                t.end()
+              }
+            })
+          )
+        )
+      })
+    })
   })
 })
